Consolidate account-related serializers and views
